move ModifiedReplay
